Refactor: 코드 리펙토링
유저 권한 설정: UserPermissionService.java
